// client/stocklinx app/src/redux/modelFieldData/constant.ts
export enum modelFieldDataConst {
  FETCH_MODELFIELDDATAS_REQUEST = "FETCH_MODELFIELDDATAS_REQUEST",
  FETCH_MODELFIELDDATAS_SUCCESS = "FETCH_MODELFIELDDATAS_SUCCESS",
  FETCH_MODELFIELDDATAS_FAILURE = "FETCH_MODELFIELDDATAS_FAILURE",
  FETCH_MODELFIELDDATA_REQUEST = "FETCH_MODELFIELDDATA_REQUEST",
  FETCH_MODELFIELDDATA_SUCCESS = "FETCH_MODELFIELDDATA_SUCCESS",
  FETCH_MODELFIELDDATA_FAILURE = "FETCH_MODELFIELDDATA_FAILURE",
  FETCH_MODELFIELDDATA_COUNTS_REQUEST = "FETCH_MODELFIELDDATA_COUNTS_REQUEST",
  FETCH_MODELFIELDDATA_COUNTS_SUCCESS = "FETCH_MODELFIELDDATA_COUNTS_SUCCESS",
  FETCH_MODELFIELDDATA_COUNTS_FAILURE = "FETCH_MODELFIELDDATA_COUNTS_FAILURE",
  CREATE_MODELFIELDDATA_REQUEST = "CREATE_MODELFIELDDATA_REQUEST",
  CREATE_MODELFIELDDATA_SUCCESS = "CREATE_MODELFIELDDATA_SUCCESS",
  CREATE_MODELFIELDDATA_FAILURE = "CREATE_MODELFIELDDATA_FAILURE",
  CREATE_RANGE_MODELFIELDDATA_REQUEST = "CREATE_RANGE_MODELFIELDDATA_REQUEST",
  CREATE_RANGE_MODELFIELDDATA_SUCCESS = "CREATE_RANGE_MODELFIELDDATA_SUCCESS",
  CREATE_RANGE_MODELFIELDDATA_FAILURE = "CREATE_RANGE_MODELFIELDDATA_FAILURE",
  UPDATE_MODELFIELDDATA_REQUEST = "UPDATE_MODELFIELDDATA_REQUEST",
  UPDATE_MODELFIELDDATA_SUCCESS = "UPDATE_MODELFIELDDATA_SUCCESS",
  UPDATE_MODELFIELDDATA_FAILURE = "UPDATE_MODELFIELDDATA_FAILURE",
  REMOVE_MODELFIELDDATA_REQUEST = "REMOVE_MODELFIELDDATA_REQUEST",
  REMOVE_MODELFIELDDATA_SUCCESS = "REMOVE_MODELFIELDDATA_SUCCESS",
  REMOVE_MODELFIELDDATA_FAILURE = "REMOVE_MODELFIELDDATA_FAILURE",
  REMOVE_RANGE_MODELFIELDDATA_REQUEST = "REMOVE_RANGE_MODELFIELDDATA_REQUEST",
  REMOVE_RANGE_MODELFIELDDATA_SUCCESS = "REMOVE_RANGE_MODELFIELDDATA_SUCCESS",
  REMOVE_RANGE_MODELFIELDDATA_FAILURE = "REMOVE_RANGE_MODELFIELDDATA_FAILURE",
  SET_MODELFIELDDATA = "SET_MODELFIELDDATA",
  CLEAR_MODELFIELDDATA = "CLEAR_MODELFIELDDATA",
  SET_MODELFIELDDATAS = "SET_MODELFIELDDATAS",
  CLEAR_MODELFIELDDATAS = "CLEAR_MODELFIELDDATAS",
}